It compiles on Linux again!

+++ src/utils/math.cpp
@@ -1,6 +1,6 @@
 #include "math.h"
 
-long double Math::dmod(long double x, long double y)
+double Math::dmod(double x, double y)
 {
     return x - floor(x/y)*y;
 }
